refactor: improve date display component architecture
